+++ src/pools/cluster/fixed.ts
@@ -102,7 +102,7 @@ export class FixedClusterPool<
 
   /** @inheritDoc */
   public get full (): boolean {
-    return this.workers.length === this.numberOfWorkers
+    return this.workerNodes.length === this.numberOfWorkers
   }
 
   /** @inheritDoc */