+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use monotonic high resolution timer for worker tasks run time.
 - Add worker tasks median run time to statistics.
+- Add worker tasks queue.
 
 ## [2.4.4] - 2023-04-07